DB change

+++ DigitalBooksApp/NewFolder/PaymentDetails.cs
@@ -5,7 +5,7 @@
         public int PaymentId { get; set; }
         public string? BuyerEmail { get; set; }
         public string? BuyerName { get; set; }
-        public long BookId { get; set; }
+        public int BookId { get; set; }
         public DateTime PaymentDate { get; set; }
     }
 }
